feature: 可注册 transition，默认使用旧的 dispatch
