+++ package.json
@@ -33,19 +33,19 @@
   "devDependencies": {
     "chai": "^3.5.0",
     "chai-as-promised": "^5.3.0",
-    "fecs": "1.4.1",
-    "fis3": "^3.4.36",
-    "fis3-hook-amd": "^0.2.0",
-    "fis3-hook-commonjs": "^0.1.25",
+    "fecs": "^1.6.0",
+    "fis3": "^3.4.39",
+    "fis3-hook-amd": "^0.2.2",
+    "fis3-hook-commonjs": "^0.1.27",
     "gitbook-cli": "^2.3.0",
-    "jsdoc-to-markdown": "^3.0.0",
+    "jsdoc-to-markdown": "^3.0.4",
     "karma": "^1.5.0",
     "karma-chai": "^0.1.0",
     "karma-chai-as-promised": "^0.1.2",
     "karma-chai-sinon": "^0.1.5",
     "karma-chrome-launcher": "^2.2.0",
-    "karma-coverage": "^1.1.1",
-    "karma-coveralls": "^1.1.2",
+    "karma-coverage": "^1.1.2",
+    "karma-coveralls": "^1.2.1",
     "karma-global-preprocessor": "0.0.2",
     "karma-html-reporter": "^0.2.7",
     "karma-mocha": "^1.1.1",
