+++ src/action.js
@@ -1,6 +1,6 @@
 /**
  * @file action.js
- * @author harttle<harttle@harttle.com>
+ * @author harttle<yangjun14@baidu.com>
  */
 
 define(function (require) {
@@ -13,6 +13,7 @@ define(function (require) {
     var assert = require('@searchfe/assert');
     var logger = require('./utils/logger');
     var _ = require('@searchfe/underscore');
+    var dispatch = require('./dispatch');
     var dom = require('./utils/dom');
     var URL = require('./utils/url');
 
@@ -27,7 +28,6 @@ define(function (require) {
 
         // The state data JUST for the next dispatch
         var stageData = {};
-        var dispatchQueue = mkDispatchQueue();
 
         function createPages() {
             var pages = cache.create('pages', {
@@ -202,50 +202,21 @@ define(function (require) {
             });
 
             doc.ensureAttached();
-            // Abort currently the running dispatch queue,
-            // and initiate a new one.
-            return dispatchQueue.reset([
-                function prevDetach() {
-                    if (!prevService) {
-                        return;
-                    }
-                    return prevService.singleton
-                        ? prevService.detach(current, prev, data)
-                        : prevService.beforeDetach(current, prev, data);
-                },
-                function currCreate() {
-                    if (!currentService) {
-                        return;
-                    }
-                    return currentService.singleton
-                        ? currentService.create(current, prev, data)
-                        : currentService.beforeAttach(current, prev, data);
-                },
-                function prevDestroy() {
-                    if (!prevService) {
-                        return;
-                    }
-                    return prevService.singleton
-                        ? prevService.destroy(current, prev, data)
-                        : prevService.detach(current, prev, data);
-                },
-                function currAttach() {
-                    if (!currentService) {
-                        return;
-                    }
-                    return currentService.attach(current, prev, data);
-                }
-            ]).exec(function currAbort() {
-                if (currentService && currentService.abort) {
-                    currentService.abort(current, prev, data);
-                }
-            }, function errorHandler(e) {
-                // eslint-disable-next-line
-                console.error(e);
-                if (_.get(current, 'options.src') !== 'sync') {
-                    location.replace(location.href);
-                }
-            });
+
+            var transitionOptions = _.assign({}, current, data);
+            var legacyArgs = {
+                current: current,
+                prev: prev,
+                data: data,
+                currentService: currentService,
+                prevService: prevService
+            };
+            return dispatch(
+                prevService,
+                currentService,
+                transitionOptions,
+                legacyArgs
+            );
         };
 
         /**
@@ -257,77 +228,6 @@ define(function (require) {
             return isIndexPage;
         };
 
-        /**
-         * Execute a queue of functions in serial, and previous execution will be stopped.
-         * This is a singleton closure containing current execution queue and threadID.
-         *
-         * A thread (implemented by mapSeries) will be initiated for each execution.
-         * And anytime there's a new thread initiating, the previous threads will stop running.
-         *
-         * @return {Object} DispatchQueue interfaces: {reset, exec}
-         * @private
-         */
-        function mkDispatchQueue() {
-            // Since we cannot quit a promise, there can be multiple threads running, actually.
-            var MAX_THREAD_COUNT = 10000;
-            // This is the ID of the currently running thread
-            var threadID = 0;
-            var lastAbortCallback;
-            var queue = [];
-            var exports = {
-                reset: reset,
-                exec: exec,
-                aborted: false
-            };
-
-            /**
-             * When reset called, a thread containing a queue of functions is initialized,
-             * and latter functions in last thread will be ommited.
-             *
-             * @param {Array} q the tasks to be queued
-             * @return {Object} The DispatchQueue object
-             */
-            function reset(q) {
-                queue = q;
-                threadID = (threadID + 1) % MAX_THREAD_COUNT;
-                return exports;
-            }
-
-            /**
-             * When exec called, current queue is executed in serial,
-             * and a promise for the results of the functions is returned.
-             *
-             * @param {Function} abortCallback The callback to be called when dispatch aborted
-             * @param {Function} errorHandler The callback to be called when error occurred
-             * @return {Promise} The promise to be resolved when all tasks completed
-             */
-            function exec(abortCallback, errorHandler) {
-                // Record the thread ID for current thread
-                // To ensure there's ONLY ONE thread running.
-                var thisThreadID = threadID;
-                if (_.isFunction(lastAbortCallback)) {
-                    lastAbortCallback();
-                }
-                lastAbortCallback = abortCallback;
-                return Promise.mapSeries(queue, function (cb) {
-                    if (typeof cb !== 'function') {
-                        return;
-                    }
-                    // Just stop running
-                    if (thisThreadID !== threadID) {
-                        return;
-                    }
-                    logger.log('calling lifecycle', cb.name || 'anonymous');
-                    return cb();
-                })
-                .catch(errorHandler)
-                .then(function () {
-                    lastAbortCallback = null;
-                });
-            }
-
-            return exports;
-        }
 
         /**
          *  Check if the specified service has been registered
@@ -368,7 +268,6 @@ define(function (require) {
          * @param {Object} data extended data being passed to `current.options`
          * */
         exports.redirect = function (url, query, options, data) {
-            var page;
             logger.log('action redirecting to: ' + url);
             exports.emit('redirecting', url);
             url = resolveUrl(url);
@@ -413,12 +312,14 @@ define(function (require) {
             page = pages.get(noRootUrl);
             if (page) {
                 page.scrollTop = window.pageYOffset;
-            } else {
+            }
+            else {
                 pages.set(noRootUrl, {
                     scrollTop: window.pageYOffset
                 });
             }
         }
+
         /**
          *  Back to last state
          *
@@ -460,6 +361,7 @@ define(function (require) {
 
             logger.log('[transfering page] from:', from, 'to:', to);
             if (!pages.contains(from)) {
+                // eslint-disable-next-line
                 console.warn('current page not found, cannot transfer to', url);
                 return;
             }
